+++ src/clientapp/src/app/roles/components/detail/role-detail.component.ts
@@ -1,9 +1,18 @@
 import { Subscription } from 'rxjs';
 
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ElementRef } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 
-import { AutoUnsubscribe, MessageBus, StatusMessage, StatusLevel } from '../../../shared';
+import {
+  AutoUnsubscribe,
+  MessageBus,
+  StatusMessage,
+  StatusLevel,
+  Popover,
+  DeleteConfirmationComponent,
+  PopoverCloseEvent,
+  DeleteConfirmation
+} from '../../../shared';
 import { RefreshMessage } from '../../../core';
 
 import { RoleDetailSlot, Role } from '../../models';
@@ -31,6 +40,8 @@ export class RoleDetailComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private service: RoleService,
+    private popup: Popover,
+    private element: ElementRef,
     private messageBus: MessageBus
   ) { }
 
@@ -66,12 +77,28 @@ export class RoleDetailComponent implements OnInit {
       return;
     }
 
-    // TODO: confirm???
+    const origin = this.element.nativeElement;
 
-    this.role$ = this.service.delete(viewModel.id)
-      .subscribe((id: string) => {
-        this.changesSaved();
-        this.back();
+    const popoverRef = this.popup
+      .open<DeleteConfirmation>({
+        content: DeleteConfirmationComponent,
+        origin,
+        hasBackdrop: false,
+        data: {
+          confirm: false,
+          itemText: viewModel.name
+        }
+      });
+
+    popoverRef.afterClosed$
+      .subscribe((res: PopoverCloseEvent<DeleteConfirmation>) => {
+        if (res.data.confirm) {
+          this.role$ = this.service.delete(viewModel.id)
+            .subscribe((id: string) => {
+              this.changesSaved();
+              this.back();
+            });
+        }
       });
   }
